main landed its own CSS box model in a93a1e9b8, independently of this branch's.
main's is the better-founded one — it was probed by round-tripping amplified
fixtures through Word's own HTML import over COM (SaveAs2 to DOCX, read the
OOXML), which refuted the padding->indent rule this branch had inferred from
crops. So HtmlParser.cs, InlineStyle.cs, docs/html-import.md and the three
HTML-CSS scenarios' baselines all resolve to main's side, and this branch's
box model is dropped.

Kept from this branch, none of which main has:

- w:lvlJc="right" right-anchors the marker, and an overflowing left-aligned
  marker pushes the first line's text to the next default tab stop
  (_probe_numtab) — MarkerRightAligned / MarkerTextShift, auto-merged clean
- HTML export text-box chrome and the missing cell-float dispatch arms
- imported list items at the bare line box with one closing block gap
  (EndListBlock), re-applied over main's parser — main still had the 4pt item
  spacing that overshot to 38.5px/row against Word's 30.5
- block edge-whitespace shedding (TrimEdgeWhitespace), also re-applied; this
  is the cause main's own html_css_margin_padding finding names

todo.md takes main's ledger, amended: #31 records the list-pitch and
edge-whitespace landings, agendas-minutes/14's fused romans and
business-plans/12's run-together section headings are deleted as closed, and
the tally is recounted from the severity lines (53/200/170 = 423 across 178 —
main's line had drifted to 55/208/170 = 433).

Baselines are NOT settled by this merge: main's for the three HTML-CSS
scenarios, this branch's for 125 others, and neither set was rendered by the
merged code. html_lists, html_nested_lists and html_css_margin_padding are
explicitly left un-re-judged. A container regeneration run is the next step.

+++ src/Morph/Fonts/GposKernTable.cs
@@ -207,7 +207,7 @@ sealed class GposKernTable
 
         if (any)
         {
-            classSubtables.Add(new([.. cov], class1, class2, (ushort) class2Count, matrix));
+            classSubtables.Add(new([.. cov], class1, class2, class2Count, matrix));
         }
     }
